Calificacion de las citas

+++ WebSite/Modelo/Paciente/CitaModelo.cs
@@ -14,7 +14,7 @@ namespace Modelo.Paciente
         public string Recomendaciones { get; set; }
         public int EstadoCita { get; set; }
         public int Calificacion { get; set; }
-        public int SesionId { get; set; }
+        public Int64 SesionId { get; set; }
         public string Dia { get; set; }
         public int Hora { get; set; }
         public string DetalleHora { get; set; }
